feat(auth): modify login logic
Instead of using Supabase's auth functionality, we're going to manually
update the users table.

The client_id from Google Analytics is inserted into the "users" table
and Supabase will generate a new user id and return it.

We'll use the new user id to set the user in our state like normal.

This logic is done when our page mounts with a useEffect with an empty
dependency array.